refactor(api): handlers_system.go 按域拆分 — 1536 行 → 5 个按域文件 (#611)
纯移动无逻辑变化：handlers_system.go(489, health/readyz/stats/protocols/features+路由注册) + handlers_auth.go(99, login/gateway 会话) + handlers_settings.go(553, 设置读写+API key 管理+存储候选) + handlers_settings_media.go(294, 流媒体/转码/自动发现设置) + handlers_sysinfo.go(189, 备份+主机指标读取+系统统计)
