+++ internal/api/handlers_sysinfo.go
@@ -1,0 +1,189 @@
+package api
+
+import (
+	"fmt"
+	"net/http"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"time"
+)
+
+func (h *Handler) handleBackup(w http.ResponseWriter, r *http.Request) {
+	if h.db == nil {
+		WriteError(w, http.StatusInternalServerError, "database not available")
+		return
+	}
+	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
+	if err := os.MkdirAll(backupDir, 0o755); err != nil {
+		WriteError(w, http.StatusInternalServerError, "failed to create backup directory")
+		return
+	}
+	filename := fmt.Sprintf("nvr-backup-%s.db", time.Now().Format("20060102-150405"))
+	destPath := filepath.Join(backupDir, filename)
+	if err := h.db.Backup(r.Context(), destPath); err != nil {
+		WriteError(w, http.StatusInternalServerError, "failed to create backup")
+		return
+	}
+	writeJSON(w, http.StatusOK, map[string]string{"status": "created", "file": filename})
+}
+
+func (h *Handler) handleListBackups(w http.ResponseWriter, r *http.Request) {
+	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
+	entries, err := os.ReadDir(backupDir)
+	if err != nil {
+		writeJSON(w, http.StatusOK, []string{})
+		return
+	}
+	var backups []string
+	for _, e := range entries {
+		if !e.IsDir() && strings.HasSuffix(e.Name(), ".db") {
+			backups = append(backups, e.Name())
+		}
+	}
+	if backups == nil {
+		backups = []string{}
+	}
+	writeJSON(w, http.StatusOK, backups)
+}
+
+// protocolInfo describes a protocol for the /api/protocols endpoint.
+
+// formatUptime converts a duration to a human-readable string like "2h 15m 30s".
+func formatUptime(d time.Duration) string {
+	rounded := d.Round(time.Second)
+	h := rounded / time.Hour
+	rounded -= h * time.Hour
+	m := rounded / time.Minute
+	rounded -= m * time.Minute
+	s := rounded / time.Second
+	if h > 0 {
+		return fmt.Sprintf("%dh %dm %ds", h, m, s)
+	}
+	if m > 0 {
+		return fmt.Sprintf("%dm %ds", m, s)
+	}
+	return fmt.Sprintf("%ds", s)
+}
+
+// --- System stats helpers (Linux /proc) ---
+
+func readCPURaw() (total, idle uint64, err error) {
+	data, err := os.ReadFile("/proc/stat")
+	if err != nil {
+		return 0, 0, err
+	}
+	lines := strings.Split(string(data), "\n")
+	if len(lines) == 0 {
+		return 0, 0, fmt.Errorf("empty /proc/stat")
+	}
+	fields := strings.Fields(lines[0])
+	if len(fields) < 5 {
+		return 0, 0, fmt.Errorf("unexpected /proc/stat format")
+	}
+	for i := 1; i < len(fields); i++ {
+		v, _ := strconv.ParseUint(fields[i], 10, 64)
+		total += v
+	}
+	idle, _ = strconv.ParseUint(fields[4], 10, 64)
+	return
+}
+
+func readMemoryInfo() (total, available uint64, err error) {
+	data, err := os.ReadFile("/proc/meminfo")
+	if err != nil {
+		return 0, 0, err
+	}
+	for _, line := range strings.Split(string(data), "\n") {
+		fields := strings.Fields(line)
+		if len(fields) < 2 {
+			continue
+		}
+		v, _ := strconv.ParseUint(fields[1], 10, 64)
+		switch fields[0] {
+		case "MemTotal:":
+			total = v * 1024
+		case "MemAvailable:":
+			available = v * 1024
+		}
+	}
+	return
+}
+
+func readNetworkInfo() (bytesSent, bytesRecv uint64, err error) {
+	data, err := os.ReadFile("/proc/net/dev")
+	if err != nil {
+		return 0, 0, err
+	}
+	// Try eth0 or wlan0 first
+	for _, line := range strings.Split(string(data), "\n") {
+		trimmed := strings.TrimSpace(line)
+		if !strings.HasPrefix(trimmed, "eth0:") && !strings.HasPrefix(trimmed, "wlan0:") {
+			continue
+		}
+		parts := strings.SplitN(trimmed, ":", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		fields := strings.Fields(parts[1])
+		if len(fields) < 10 {
+			continue
+		}
+		bytesRecv, _ = strconv.ParseUint(fields[0], 10, 64)
+		bytesSent, _ = strconv.ParseUint(fields[8], 10, 64)
+		return
+	}
+	// Fallback: sum all interfaces
+	for _, line := range strings.Split(string(data), "\n") {
+		trimmed := strings.TrimSpace(line)
+		if !strings.Contains(trimmed, ":") {
+			continue
+		}
+		parts := strings.SplitN(trimmed, ":", 2)
+		if len(parts) < 2 {
+			continue
+		}
+		fields := strings.Fields(parts[1])
+		if len(fields) < 10 {
+			continue
+		}
+		r, _ := strconv.ParseUint(fields[0], 10, 64)
+		s, _ := strconv.ParseUint(fields[8], 10, 64)
+		bytesRecv += r
+		bytesSent += s
+	}
+	return
+}
+
+func readProcessRSS() uint64 {
+	data, err := os.ReadFile("/proc/self/statm")
+	if err != nil {
+		return 0
+	}
+	fields := strings.Fields(string(data))
+	if len(fields) < 2 {
+		return 0
+	}
+	rssPages, _ := strconv.ParseUint(fields[1], 10, 64)
+	return rssPages * uint64(os.Getpagesize())
+}
+
+func (h *Handler) handleSystemStats(w http.ResponseWriter, r *http.Request) {
+	cpuTotal, cpuIdle, _ := readCPURaw()
+	memTotal, memAvailable, _ := readMemoryInfo()
+	netSent, netRecv, _ := readNetworkInfo()
+	processRSS := readProcessRSS()
+
+	writeJSON(w, http.StatusOK, SystemStats{
+		CPU:       CPUStats{Total: cpuTotal, Idle: cpuIdle},
+		Memory:    MemoryStats{Total: memTotal, Available: memAvailable, ProcessRSS: processRSS},
+		Network:   NetworkStats{BytesSent: netSent, BytesRecv: netRecv},
+		Uptime:    formatUptime(time.Since(appStartTime)),
+		Timestamp: time.Now().Unix(),
+	})
+}
+
+// handleGetAutoDiscoverSettings returns the current auto_discover config. The
+// default_password is NEVER returned over the API (avoid plaintext leakage);
+// instead a boolean has_default_password indicates whether one is configured.
