+++ internal/api/handlers_system.go
@@ -5,24 +5,17 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/http/httptest"
-	"os"
-	"path/filepath"
 	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/config"
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/middleware"
-	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/storage"
 
 	"github.com/go-chi/chi/v5"
 
 	"github.com/Mi-Bee-Studio/MiBeeNvr/internal/model"
 )
-
-// --- Public endpoints ---
 
 func (h *Handler) handleHealth(w http.ResponseWriter, r *http.Request) {
 	resp := HealthResponse{Checks: make(map[string]HealthCheck)}
@@ -143,6 +136,8 @@ func (h *Handler) handleHealth(w http.ResponseWriter, r *http.Request) {
 }
 
 // aggregateCameraHealth builds a CameraHealthSummary from the health manager and camera DB.
+
+// aggregateCameraHealth builds a CameraHealthSummary from the health manager and camera DB.
 func (h *Handler) aggregateCameraHealth(r *http.Request) *CameraHealthSummary {
 	allHealth := h.healthMgr.GetAllHealth()
 	if allHealth == nil {
@@ -196,6 +191,9 @@ func (h *Handler) aggregateCameraHealth(r *http.Request) *CameraHealthSummary {
 
 	return summary
 }
+
+// handleHealthCameras returns full camera health map with scores.
+// Public endpoint — no auth required.
 
 // handleHealthCameras returns full camera health map with scores.
 // Public endpoint — no auth required.
@@ -269,89 +267,6 @@ func (h *Handler) handleReadyz(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func (h *Handler) handleLogin(w http.ResponseWriter, r *http.Request) {
-	// Validate credentials by running through the auth middleware.
-	// If auth is disabled, any request succeeds; otherwise BasicAuth is checked.
-	// Use httptest.ResponseRecorder to capture middleware output without writing to client w.
-	done := make(chan int, 1)
-
-	inner := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		done <- http.StatusOK
-	})
-	rec := httptest.NewRecorder()
-	h.authMW(inner).ServeHTTP(rec, r)
-
-	select {
-	case status := <-done:
-		if status == http.StatusOK {
-			// Credentials validated by the middleware (BasicAuth path). Mint a
-			// stateless signed session token so the browser can stop carrying the
-			// reversible base64(user:pass). The username comes from the request's
-			// BasicAuth header (just validated); the bcrypt hash from config drives
-			// the HMAC key, so a later password change invalidates this token.
-			// Tests use a nil-config handler, so guard against that here.
-			if h.config == nil {
-				writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
-				return
-			}
-			username := h.config.Auth.Username
-			if u, _, ok := r.BasicAuth(); ok && u != "" {
-				username = u
-			}
-			hash := h.config.Auth.PasswordHash
-			token, expiresAt := middleware.SignSessionToken(username, hash, time.Now())
-			writeJSON(w, http.StatusOK, map[string]string{
-				"status":     "ok",
-				"token":      token,
-				"expires_at": expiresAt.UTC().Format(time.RFC3339),
-			})
-		}
-	default:
-		// Forward the middleware's captured response (503 SETUP_REQUIRED, 401, etc.)
-		// without double-writing to the client.
-		for k, vv := range rec.Header() {
-			for _, v := range vv {
-				w.Header().Add(k, v)
-			}
-		}
-		w.WriteHeader(rec.Code)
-		w.Write(rec.Body.Bytes())
-	}
-}
-
-// handleGatewaySession mints an NVR session token for a request that arrived
-// through the fnOS unified gateway with a verified ADMIN identity (issue #394).
-// The SPA calls this on boot when it has no stored token: inside the fnOS
-// desktop (gateway iframe) it gets a token and skips the login page; accessed
-// directly on :9090 there is no gateway identity context, so it always 401s
-// and the SPA shows the normal login form.
-func (h *Handler) handleGatewaySession(w http.ResponseWriter, r *http.Request) {
-	gi := middleware.GatewayIdentityFromContext(r.Context())
-	if gi == nil || !gi.Admin {
-		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "not available on this listener"})
-		return
-	}
-	if h.config == nil {
-		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "no config"})
-		return
-	}
-	hash := h.config.Auth.PasswordHash
-	if hash == "" {
-		w.Header().Set("Content-Type", "application/json")
-		w.Header().Set("WWW-Authenticate", `Basic realm="MiBee NVR"`)
-		w.WriteHeader(http.StatusServiceUnavailable)
-		w.Write([]byte(`{"error":"setup required","code":"SETUP_REQUIRED"}`))
-		return
-	}
-	token, expiresAt := middleware.SignSessionToken(h.config.Auth.Username, hash, time.Now())
-	writeJSON(w, http.StatusOK, map[string]string{
-		"status":       "ok",
-		"token":        token,
-		"expires_at":   expiresAt.UTC().Format(time.RFC3339),
-		"gateway_user": gi.Username,
-	})
-}
-
 func (h *Handler) handleStats(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
@@ -415,755 +330,6 @@ func (h *Handler) handleStatsTrends(w http.ResponseWriter, r *http.Request) {
 }
 
 // --- Settings endpoints ---
-
-func (h *Handler) handleGetSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	// Compute timezone display string
-	tzDisplay := h.config.Timezone
-	if tzDisplay != "" && tzDisplay != "UTC" && tzDisplay != "Local" {
-		if loc, err := time.LoadLocation(tzDisplay); err == nil {
-			_, offset := time.Now().In(loc).Zone()
-			tzDisplay = fmt.Sprintf("%s (UTC%s)", tzDisplay, formatOffset(offset))
-		}
-	} else if tzDisplay == "UTC" {
-		tzDisplay = "UTC"
-	} else if tzDisplay == "Local" {
-		if loc, err := time.LoadLocation("Local"); err == nil {
-			name, offset := time.Now().In(loc).Zone()
-			tzDisplay = fmt.Sprintf("%s (UTC%s)", name, formatOffset(offset))
-		} else {
-			tzDisplay = "Local"
-		}
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"cleanup": map[string]any{
-			"retention_days":         h.config.Cleanup.RetentionDays,
-			"check_interval":         h.config.Cleanup.CheckInterval,
-			"disk_threshold_percent": h.config.Cleanup.DiskThresholdPercent,
-			// Motion-aware disk cleanup (#435): when the disk threshold is
-			// hit, evict boring (low motion_score) segments first. nil = on.
-			"motion_aware_disk_cleanup": h.config.Cleanup.MotionAwareDiskCleanup == nil || *h.config.Cleanup.MotionAwareDiskCleanup,
-		},
-		"webdav": map[string]any{
-			"enabled":     h.config.WebDAV.Enabled != nil && *h.config.WebDAV.Enabled,
-			"path_prefix": h.config.WebDAV.PathPrefix,
-			"read_write":  h.config.WebDAV.ReadWrite,
-		},
-		"storage": map[string]any{
-			"root_dir": h.config.Storage.RootDir,
-		},
-		"auth": map[string]any{
-			"username":        h.config.Auth.Username,
-			"auth_configured": h.config.Auth.PasswordHash != "" || h.config.Auth.Password != "",
-		},
-		"mibeevision": map[string]any{
-			"api_keys": buildAPIKeyInfo(h.config.APIKeys, h.apiKeyLastUsed()),
-		},
-		"timezone":         h.config.Timezone,
-		"timezone_display": tzDisplay,
-		"server": map[string]any{
-			"listen": h.config.Server.Listen,
-		},
-		"gb28181": map[string]any{
-			"enabled":    h.config.GB28181.Enabled,
-			"sip_listen": h.config.GB28181.SIPListen,
-			"server_id":  h.config.GB28181.ServerID,
-			"realm":      h.config.GB28181.Realm,
-			// Never return the SIP password (mirrors the masked auth/API-key
-			// fields above); the frontend only needs whether one is set.
-			"password_configured": h.config.GB28181.Password != "",
-			"port_range":          h.config.GB28181.PortRange,
-			"allowed_device_ids":  h.config.GB28181.AllowedDeviceIDs,
-			"heartbeat_interval":  h.config.GB28181.HeartbeatInterval,
-			"catalog_interval":    h.config.GB28181.CatalogInterval,
-			"tcp_mode":            h.config.GB28181.TCPMode,
-			"tcp_framing":         h.config.GB28181.TCPFraming,
-			"media_transport":     h.config.GB28181.MediaTransport,
-			"sip_transport":       h.config.GB28181.SIPTransport,
-			// Subscription toggles (#341): resolve the *bool defaults so the
-			// UI sees the effective values.
-			"subscribe_catalog":         h.config.GB28181.CatalogSubscriptionOn(),
-			"subscribe_alarm":           h.config.GB28181.AlarmSubscriptionOn(),
-			"subscribe_mobile_position": h.config.GB28181.SubscribeMobilePosition,
-			"subscribe_expires":         h.config.GB28181.SubscribeExpires,
-		},
-	})
-}
-
-func formatOffset(seconds int) string {
-	sign := "+"
-	if seconds < 0 {
-		sign = "-"
-		seconds = -seconds
-	}
-	hours := seconds / 3600
-	mins := (seconds % 3600) / 60
-	if mins == 0 {
-		return fmt.Sprintf("%s%d", sign, hours)
-	}
-	return fmt.Sprintf("%s%d:%02d", sign, hours, mins)
-}
-
-// buildAPIKeyInfo returns a safe summary of configured API keys (never the key
-// itself). Revoked keys are included (grayed in the UI) so owners can see the
-// full per-device token list; last_used comes from the live store when wired.
-func buildAPIKeyInfo(keys []config.APIKeyConfig, lastUsed map[string]time.Time) []map[string]any {
-	result := make([]map[string]any, 0, len(keys))
-	for _, k := range keys {
-		info := map[string]any{
-			"name":    k.Name,
-			"prefix":  k.Key[:min(8, len(k.Key))] + "…", // e.g. "mbv_ab12…"
-			"revoked": k.Revoked,
-		}
-		if t, ok := lastUsed[k.Name]; ok && !t.IsZero() {
-			info["last_used"] = t.UTC().Format(time.RFC3339)
-		}
-		result = append(result, info)
-	}
-	return result
-}
-
-// syncAPIKeyStore rebuilds the live key set from the in-memory config. Called
-// after generate/revoke mutations so the API-key middleware picks up the
-// change on the next request without a restart (#335).
-func (h *Handler) syncAPIKeyStore() {
-	if h.apiKeyStore == nil || h.config == nil {
-		return
-	}
-	valid := make(map[string]string)
-	for _, k := range h.config.APIKeys {
-		if !k.Revoked && k.Key != "" {
-			valid[k.Key] = k.Name
-		}
-	}
-	h.apiKeyStore.SetKeys(valid)
-}
-
-// apiKeyLastUsed returns the store's last-used map, or nil when no store is
-// wired (tests, sub-handlers).
-func (h *Handler) apiKeyLastUsed() map[string]time.Time {
-	if h.apiKeyStore == nil {
-		return nil
-	}
-	return h.apiKeyStore.LastUsed()
-}
-
-func (h *Handler) handleUpdateSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	var body struct {
-		Cleanup *struct {
-			RetentionDays        *int    `json:"retention_days"`
-			DiskThresholdPercent *int    `json:"disk_threshold_percent"`
-			CheckInterval        *string `json:"check_interval"`
-			// Motion-aware disk cleanup toggle (#435): evict boring segments
-			// first when the disk threshold is hit. nil = unchanged.
-			MotionAwareDiskCleanup *bool `json:"motion_aware_disk_cleanup"`
-		} `json:"cleanup"`
-		Storage *struct {
-			// Recording root directory (#395). Takes effect on the NEXT start —
-			// the DB and all subsystems open paths under the old root at boot,
-			// so a live switch would corrupt state. The response carries
-			// restart_required=true when this changes.
-			RootDir *string `json:"root_dir"`
-		} `json:"storage"`
-		WebDAV *struct {
-			Enabled    *bool   `json:"enabled"`
-			PathPrefix *string `json:"path_prefix"`
-			ReadWrite  *bool   `json:"read_write"`
-		} `json:"webdav"`
-		Timezone *string `json:"timezone"`
-		Server   *struct {
-			Listen *string `json:"listen"`
-		} `json:"server"`
-		GB28181 *struct {
-			Enabled           *bool     `json:"enabled"`
-			SIPListen         *string   `json:"sip_listen"`
-			ServerID          *string   `json:"server_id"`
-			Realm             *string   `json:"realm"`
-			Password          *string   `json:"password"`
-			PortRange         *string   `json:"port_range"`
-			AllowedDeviceIDs  *[]string `json:"allowed_device_ids"`
-			HeartbeatInterval *string   `json:"heartbeat_interval"`
-			CatalogInterval   *string   `json:"catalog_interval"`
-			TCPMode           *bool     `json:"tcp_mode"`
-			TCPFraming        *string   `json:"tcp_framing"`
-			MediaTransport    *string   `json:"media_transport"`
-			SIPTransport      *string   `json:"sip_transport"`
-
-			SubscribeCatalog        *bool   `json:"subscribe_catalog"`
-			SubscribeAlarm          *bool   `json:"subscribe_alarm"`
-			SubscribeMobilePosition *bool   `json:"subscribe_mobile_position"`
-			SubscribeExpires        *string `json:"subscribe_expires"`
-		} `json:"gb28181"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	// Update cleanup settings
-	if body.Cleanup != nil {
-		if body.Cleanup.RetentionDays != nil {
-			if *body.Cleanup.RetentionDays < 1 {
-				WriteError(w, http.StatusBadRequest, "retention_days must be >= 1")
-				return
-			}
-			h.config.Cleanup.RetentionDays = *body.Cleanup.RetentionDays
-		}
-		if body.Cleanup.DiskThresholdPercent != nil {
-			if *body.Cleanup.DiskThresholdPercent < 1 || *body.Cleanup.DiskThresholdPercent > 100 {
-				WriteError(w, http.StatusBadRequest, "disk_threshold_percent must be between 1 and 100")
-				return
-			}
-			h.config.Cleanup.DiskThresholdPercent = *body.Cleanup.DiskThresholdPercent
-		}
-		if body.Cleanup.CheckInterval != nil {
-			// An empty/whitespace string means "keep current value" (partial PUT
-			// semantics). Previously any non-nil value — including "" sent by the
-			// cleanup settings UI — hit time.ParseDuration("") → 400 and aborted
-			// the whole cleanup save (#294). Treat blank as "no change".
-			if trimmed := strings.TrimSpace(*body.Cleanup.CheckInterval); trimmed != "" {
-				if _, err := time.ParseDuration(trimmed); err != nil {
-					WriteError(w, http.StatusBadRequest, "check_interval must be a valid duration (e.g., \"30m\", \"1h\")")
-					return
-				}
-				h.config.Cleanup.CheckInterval = trimmed
-			}
-		}
-		if body.Cleanup.MotionAwareDiskCleanup != nil {
-			h.config.Cleanup.MotionAwareDiskCleanup = body.Cleanup.MotionAwareDiskCleanup
-		}
-	}
-
-	// Update storage root (#395) — next-start semantics, see body.Storage.
-	storageChanged := false
-	if body.Storage != nil && body.Storage.RootDir != nil {
-		dir := strings.TrimSpace(*body.Storage.RootDir)
-		if dir == "" || !strings.HasPrefix(dir, "/") {
-			WriteError(w, http.StatusBadRequest, "storage.root_dir must be an absolute path")
-			return
-		}
-		dir = strings.TrimRight(dir, "/")
-		if dir == "" {
-			dir = "/"
-		}
-		if dir != h.config.Storage.RootDir {
-			// Preflight the target with a real SQLite round-trip (create dir,
-			// create+drop a table under the same WAL/mmap pragmas the NVR DB
-			// uses). MkdirAll alone accepted paths that later crash-looped the
-			// app at db init — some platforms (fnOS external storage) take
-			// plain file creation but reject the syscalls SQLite needs.
-			if err := storage.ProbeRoot(dir); err != nil {
-				WriteError(w, http.StatusBadRequest,
-					fmt.Sprintf("storage root not usable: %v (the recording root must host the NVR database)", err))
-				return
-			}
-			h.config.Storage.RootDir = dir
-			// Hot switch: NEW segments immediately go to the new root (the DB
-			// is decoupled and stays on the data volume; in-flight segments
-			// finish where they started). No restart needed.
-			if h.store != nil {
-				if err := h.store.SetRootDir(dir); err != nil {
-					WriteError(w, http.StatusBadRequest, fmt.Sprintf("cannot switch recording root: %v", err))
-					return
-				}
-			}
-			storageChanged = true
-			logger.Info("storage root_dir switched (hot)", "new_root", dir)
-		}
-	}
-
-	// Update webdav settings
-	if body.WebDAV != nil {
-		if body.WebDAV.Enabled != nil {
-			if h.config.WebDAV.Enabled == nil {
-				h.config.WebDAV.Enabled = new(bool)
-			}
-			*h.config.WebDAV.Enabled = *body.WebDAV.Enabled
-		}
-		if body.WebDAV.PathPrefix != nil {
-			h.config.WebDAV.PathPrefix = *body.WebDAV.PathPrefix
-		}
-		if body.WebDAV.ReadWrite != nil {
-			h.config.WebDAV.ReadWrite = *body.WebDAV.ReadWrite
-		}
-	}
-
-	// Update timezone
-	if body.Timezone != nil {
-		tz := strings.TrimSpace(*body.Timezone)
-		if tz != "" && tz != "UTC" && tz != "Local" {
-			if _, err := time.LoadLocation(tz); err != nil {
-				WriteError(w, http.StatusBadRequest, fmt.Sprintf("invalid timezone: %q", tz))
-				return
-			}
-		}
-		h.config.Timezone = tz
-	}
-
-	// Update server listen port
-	if body.Server != nil && body.Server.Listen != nil {
-		raw := strings.TrimSpace(*body.Server.Listen)
-		raw = strings.TrimPrefix(raw, ":")
-		port, err := strconv.Atoi(raw)
-		if err != nil || port < 1 || port > 65535 {
-			WriteError(w, http.StatusBadRequest, "listen must be a valid port (1-65535)")
-			return
-		}
-		h.config.Server.Listen = fmt.Sprintf(":%d", port)
-	}
-
-	// Update GB28181 settings
-	if body.GB28181 != nil {
-		if body.GB28181.Enabled != nil {
-			h.config.GB28181.Enabled = *body.GB28181.Enabled
-		}
-		if body.GB28181.SIPListen != nil {
-			h.config.GB28181.SIPListen = *body.GB28181.SIPListen
-		}
-		if body.GB28181.ServerID != nil {
-			h.config.GB28181.ServerID = *body.GB28181.ServerID
-		}
-		if body.GB28181.Realm != nil {
-			h.config.GB28181.Realm = *body.GB28181.Realm
-		}
-		if body.GB28181.Password != nil {
-			// Blank password = keep current (the GET no longer returns the
-			// password, so the UI round-trips an empty field when unchanged).
-			if trimmed := strings.TrimSpace(*body.GB28181.Password); trimmed != "" {
-				h.config.GB28181.Password = trimmed
-			}
-		}
-		if body.GB28181.PortRange != nil {
-			h.config.GB28181.PortRange = *body.GB28181.PortRange
-		}
-		if body.GB28181.AllowedDeviceIDs != nil {
-			h.config.GB28181.AllowedDeviceIDs = *body.GB28181.AllowedDeviceIDs
-		}
-		if body.GB28181.HeartbeatInterval != nil {
-			h.config.GB28181.HeartbeatInterval = *body.GB28181.HeartbeatInterval
-		}
-		if body.GB28181.CatalogInterval != nil {
-			h.config.GB28181.CatalogInterval = *body.GB28181.CatalogInterval
-		}
-		if body.GB28181.TCPMode != nil {
-			h.config.GB28181.TCPMode = *body.GB28181.TCPMode
-		}
-		if body.GB28181.TCPFraming != nil {
-			h.config.GB28181.TCPFraming = *body.GB28181.TCPFraming
-		}
-		if body.GB28181.MediaTransport != nil {
-			h.config.GB28181.MediaTransport = *body.GB28181.MediaTransport
-			// Keep the legacy alias coherent for older config readers.
-			h.config.GB28181.TCPMode = *body.GB28181.MediaTransport != "udp"
-		}
-		if body.GB28181.SIPTransport != nil {
-			h.config.GB28181.SIPTransport = *body.GB28181.SIPTransport
-		}
-		if body.GB28181.SubscribeCatalog != nil {
-			h.config.GB28181.SubscribeCatalog = body.GB28181.SubscribeCatalog
-		}
-		if body.GB28181.SubscribeAlarm != nil {
-			h.config.GB28181.SubscribeAlarm = body.GB28181.SubscribeAlarm
-		}
-		if body.GB28181.SubscribeMobilePosition != nil {
-			h.config.GB28181.SubscribeMobilePosition = *body.GB28181.SubscribeMobilePosition
-		}
-		if body.GB28181.SubscribeExpires != nil {
-			if trimmed := strings.TrimSpace(*body.GB28181.SubscribeExpires); trimmed != "" {
-				if _, err := time.ParseDuration(trimmed); err != nil {
-					WriteError(w, http.StatusBadRequest, "subscribe_expires must be a valid duration (e.g., \"3600s\", \"2h\")")
-					return
-				}
-				h.config.GB28181.SubscribeExpires = trimmed
-			}
-		}
-		// Validate the updated config (catches invalid server_id, sip_listen, etc.)
-		if err := config.Validate(h.config); err != nil {
-			WriteError(w, http.StatusBadRequest, err.Error())
-			return
-		}
-	}
-
-	// Persist config to disk
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-	}
-
-	if storageChanged {
-		writeJSON(w, http.StatusOK, map[string]any{"status": "updated", "restart_required": false})
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-// handleStorageCandidates reports the recording-root choices available to the
-// settings UI (#395): the current storage.root_dir plus any extra locations the
-// host platform granted (fnOS user-authorized dirs, mounted under /media/* by
-// the lifecycle script and passed via NVR_STORAGE_CANDIDATES).
-func (h *Handler) handleStorageCandidates(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-	type candidate struct {
-		Path  string `json:"path"`
-		Label string `json:"label"`
-	}
-	resp := struct {
-		Current     string      `json:"current"`
-		Candidates  []candidate `json:"candidates"`
-		RestartHint string      `json:"restart_hint"`
-		EnvManaged  bool        `json:"env_managed"`
-	}{
-		Current:     h.config.Storage.RootDir,
-		Candidates:  []candidate{{Path: h.config.Storage.RootDir, Label: "current"}},
-		RestartHint: "切换立即生效：新录像将写入所选位置（无需重启）",
-		EnvManaged:  os.Getenv("NVR_STORAGE_CANDIDATES") != "",
-	}
-	seen := map[string]bool{h.config.Storage.RootDir: true}
-	for _, p := range h.config.Storage.Candidates {
-		if p == "" || seen[p] {
-			continue
-		}
-		// Live existence filter: a mount revoked or unplugged since boot must
-		// not keep offering a dead choice here (the startup-time prune in
-		// config.ApplyDefaults only covers the persisted list).
-		if info, err := os.Stat(p); err != nil || !info.IsDir() {
-			continue
-		}
-		seen[p] = true
-		label := strings.TrimPrefix(p, "/media/")
-		if label == p {
-			label = filepath.Base(p)
-		}
-		resp.Candidates = append(resp.Candidates, candidate{Path: p, Label: label})
-	}
-	writeJSON(w, http.StatusOK, resp)
-}
-
-// handleGenerateAPIKey creates a new API key for MiBeeVision integration.
-// POST /api/settings/api-keys  body: {"name": "mibeevision-prod"}
-// Returns the full key ONCE (never exposed again).
-func (h *Handler) handleGenerateAPIKey(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-	var body struct {
-		Name string `json:"name"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	name := strings.TrimSpace(body.Name)
-	if name == "" {
-		name = "mibeevision"
-	}
-
-	key := middleware.GenerateAPIKey()
-	h.config.APIKeys = append(h.config.APIKeys, config.APIKeyConfig{
-		Key:  key,
-		Name: name,
-	})
-
-	if err := config.Save(h.configPath, h.config); err != nil {
-		WriteError(w, http.StatusInternalServerError, "failed to save config")
-		return
-	}
-	h.syncAPIKeyStore()
-
-	logger.Info("API key generated", "name", name)
-	writeJSON(w, http.StatusCreated, map[string]any{
-		"name":   name,
-		"key":    key,
-		"prefix": key[:min(8, len(key))] + "…",
-	})
-}
-
-// handleRevokeAPIKey marks an API key as revoked by name.
-// DELETE /api/settings/api-keys/{name}
-func (h *Handler) handleRevokeAPIKey(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-	name := chi.URLParam(r, "name")
-	if name == "" {
-		WriteError(w, http.StatusBadRequest, "name is required")
-		return
-	}
-
-	found := false
-	for i := range h.config.APIKeys {
-		if h.config.APIKeys[i].Name == name {
-			h.config.APIKeys[i].Revoked = true
-			found = true
-			break
-		}
-	}
-	if !found {
-		WriteError(w, http.StatusNotFound, "API key not found")
-		return
-	}
-
-	if err := config.Save(h.configPath, h.config); err != nil {
-		WriteError(w, http.StatusInternalServerError, "failed to save config")
-		return
-	}
-	h.syncAPIKeyStore()
-
-	logger.Info("API key revoked", "name", name)
-	writeJSON(w, http.StatusOK, map[string]string{"status": "revoked"})
-}
-
-func (h *Handler) handleGetStreamingSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"webrtc": map[string]any{
-			"enabled":      h.config.Streaming.WebRTC.Enabled != nil && *h.config.Streaming.WebRTC.Enabled,
-			"max_viewers":  h.config.Streaming.WebRTC.MaxViewers,
-			"idle_timeout": h.config.Streaming.WebRTC.IdleTimeout,
-		},
-		"flv": map[string]any{
-			"enabled":        h.config.Streaming.FLV.Enabled != nil && *h.config.Streaming.FLV.Enabled,
-			"max_viewers":    h.config.Streaming.FLV.MaxViewers,
-			"idle_timeout":   h.config.Streaming.FLV.IdleTimeout,
-			"gop_cache_size": h.config.Streaming.FLV.GOPCacheSize,
-		},
-		"hls": map[string]any{
-			"low_latency": h.config.HLS.LowLatency,
-		},
-		// RTMP/SRT ingest settings. The streaming panel persisted nothing for
-		// them before: the PUT body had no rtmp/srt fields, so the UI switches
-		// silently no-op'd while still toasting success (found configuring the
-		// fnOS test box, 2026-08-19).
-		"rtmp": map[string]any{
-			"enabled":     h.config.RTMP.Enabled != nil && *h.config.RTMP.Enabled,
-			"port":        h.config.RTMP.Port,
-			"stream_keys": h.config.RTMP.StreamKeys, // camera_id → stream_key (legacy map; per-camera stream_key takes precedence)
-		},
-		"srt": map[string]any{
-			"enabled": h.config.SRT.Enabled != nil && *h.config.SRT.Enabled,
-			"port":    h.config.SRT.Port,
-			"streams": h.config.SRT.Streams,
-		},
-	})
-}
-
-func (h *Handler) handleUpdateStreamingSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	var body struct {
-		WebRTC *struct {
-			Enabled     *bool   `json:"enabled"`
-			MaxViewers  *int    `json:"max_viewers"`
-			IdleTimeout *string `json:"idle_timeout"`
-		} `json:"webrtc"`
-		FLV *struct {
-			Enabled      *bool   `json:"enabled"`
-			MaxViewers   *int    `json:"max_viewers"`
-			IdleTimeout  *string `json:"idle_timeout"`
-			GOPCacheSize *int    `json:"gop_cache_size"`
-		} `json:"flv"`
-		HLS *struct {
-			LowLatency *bool `json:"low_latency"`
-		} `json:"hls"`
-		RTMP *struct {
-			Enabled    *bool             `json:"enabled"`
-			Port       *int              `json:"port"`
-			StreamKeys map[string]string `json:"stream_keys"` // camera_id → stream_key
-		} `json:"rtmp"`
-		SRT *struct {
-			Enabled *bool              `json:"enabled"`
-			Port    *int               `json:"port"`
-			Streams []config.SRTStream `json:"streams"`
-		} `json:"srt"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	if body.WebRTC != nil {
-		if body.WebRTC.Enabled != nil {
-			if h.config.Streaming.WebRTC.Enabled == nil {
-				h.config.Streaming.WebRTC.Enabled = new(bool)
-			}
-			*h.config.Streaming.WebRTC.Enabled = *body.WebRTC.Enabled
-		}
-		if body.WebRTC.MaxViewers != nil {
-			h.config.Streaming.WebRTC.MaxViewers = *body.WebRTC.MaxViewers
-		}
-		if body.WebRTC.IdleTimeout != nil {
-			h.config.Streaming.WebRTC.IdleTimeout = *body.WebRTC.IdleTimeout
-		}
-	}
-
-	if body.FLV != nil {
-		if body.FLV.Enabled != nil {
-			if h.config.Streaming.FLV.Enabled == nil {
-				h.config.Streaming.FLV.Enabled = new(bool)
-			}
-			*h.config.Streaming.FLV.Enabled = *body.FLV.Enabled
-		}
-		if body.FLV.MaxViewers != nil {
-			h.config.Streaming.FLV.MaxViewers = *body.FLV.MaxViewers
-		}
-		if body.FLV.IdleTimeout != nil {
-			h.config.Streaming.FLV.IdleTimeout = *body.FLV.IdleTimeout
-		}
-		if body.FLV.GOPCacheSize != nil {
-			h.config.Streaming.FLV.GOPCacheSize = *body.FLV.GOPCacheSize
-		}
-	}
-
-	if body.HLS != nil && body.HLS.LowLatency != nil {
-		h.config.HLS.LowLatency = *body.HLS.LowLatency
-	}
-
-	if body.RTMP != nil {
-		if body.RTMP.Enabled != nil {
-			if h.config.RTMP.Enabled == nil {
-				h.config.RTMP.Enabled = new(bool)
-			}
-			*h.config.RTMP.Enabled = *body.RTMP.Enabled
-		}
-		if body.RTMP.Port != nil {
-			h.config.RTMP.Port = *body.RTMP.Port
-		}
-		if body.RTMP.StreamKeys != nil {
-			h.config.RTMP.StreamKeys = body.RTMP.StreamKeys
-		}
-	}
-
-	if body.SRT != nil {
-		if body.SRT.Enabled != nil {
-			if h.config.SRT.Enabled == nil {
-				h.config.SRT.Enabled = new(bool)
-			}
-			*h.config.SRT.Enabled = *body.SRT.Enabled
-		}
-		if body.SRT.Port != nil {
-			h.config.SRT.Port = *body.SRT.Port
-		}
-		if body.SRT.Streams != nil {
-			h.config.SRT.Streams = body.SRT.Streams
-		}
-	}
-
-	// Persist config to disk
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-func (h *Handler) handleGetTranscodingSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	writeJSON(w, http.StatusOK, map[string]any{
-		"enabled":     h.config.Transcoding.Enabled,
-		"max_workers": h.config.Transcoding.MaxWorkers,
-	})
-}
-
-func (h *Handler) handleUpdateTranscodingSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-
-	var body struct {
-		Enabled    *bool `json:"enabled"`
-		MaxWorkers *int  `json:"max_workers"`
-	}
-
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	if body.MaxWorkers != nil {
-		if *body.MaxWorkers < 1 || *body.MaxWorkers > 4 {
-			WriteError(w, http.StatusBadRequest, "max_workers must be between 1 and 4")
-			return
-		}
-		h.config.Transcoding.MaxWorkers = *body.MaxWorkers
-	}
-
-	if body.Enabled != nil {
-		h.config.Transcoding.Enabled = *body.Enabled
-	}
-
-	// Persist config to disk
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-	}
-
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
-
-func (h *Handler) handleBackup(w http.ResponseWriter, r *http.Request) {
-	if h.db == nil {
-		WriteError(w, http.StatusInternalServerError, "database not available")
-		return
-	}
-	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
-	if err := os.MkdirAll(backupDir, 0o755); err != nil {
-		WriteError(w, http.StatusInternalServerError, "failed to create backup directory")
-		return
-	}
-	filename := fmt.Sprintf("nvr-backup-%s.db", time.Now().Format("20060102-150405"))
-	destPath := filepath.Join(backupDir, filename)
-	if err := h.db.Backup(r.Context(), destPath); err != nil {
-		WriteError(w, http.StatusInternalServerError, "failed to create backup")
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "created", "file": filename})
-}
-
-func (h *Handler) handleListBackups(w http.ResponseWriter, r *http.Request) {
-	backupDir := filepath.Join(filepath.Dir(h.configPath), "backups")
-	entries, err := os.ReadDir(backupDir)
-	if err != nil {
-		writeJSON(w, http.StatusOK, []string{})
-		return
-	}
-	var backups []string
-	for _, e := range entries {
-		if !e.IsDir() && strings.HasSuffix(e.Name(), ".db") {
-			backups = append(backups, e.Name())
-		}
-	}
-	if backups == nil {
-		backups = []string{}
-	}
-	writeJSON(w, http.StatusOK, backups)
-}
 
 // protocolInfo describes a protocol for the /api/protocols endpoint.
 type protocolInfo struct {
@@ -1287,219 +453,6 @@ func (h *Handler) handleUpdateFeatures(w http.ResponseWriter, r *http.Request) {
 }
 
 // formatUptime converts a duration to a human-readable string like "2h 15m 30s".
-func formatUptime(d time.Duration) string {
-	rounded := d.Round(time.Second)
-	h := rounded / time.Hour
-	rounded -= h * time.Hour
-	m := rounded / time.Minute
-	rounded -= m * time.Minute
-	s := rounded / time.Second
-	if h > 0 {
-		return fmt.Sprintf("%dh %dm %ds", h, m, s)
-	}
-	if m > 0 {
-		return fmt.Sprintf("%dm %ds", m, s)
-	}
-	return fmt.Sprintf("%ds", s)
-}
-
-// --- System stats helpers (Linux /proc) ---
-
-func readCPURaw() (total, idle uint64, err error) {
-	data, err := os.ReadFile("/proc/stat")
-	if err != nil {
-		return 0, 0, err
-	}
-	lines := strings.Split(string(data), "\n")
-	if len(lines) == 0 {
-		return 0, 0, fmt.Errorf("empty /proc/stat")
-	}
-	fields := strings.Fields(lines[0])
-	if len(fields) < 5 {
-		return 0, 0, fmt.Errorf("unexpected /proc/stat format")
-	}
-	for i := 1; i < len(fields); i++ {
-		v, _ := strconv.ParseUint(fields[i], 10, 64)
-		total += v
-	}
-	idle, _ = strconv.ParseUint(fields[4], 10, 64)
-	return
-}
-
-func readMemoryInfo() (total, available uint64, err error) {
-	data, err := os.ReadFile("/proc/meminfo")
-	if err != nil {
-		return 0, 0, err
-	}
-	for _, line := range strings.Split(string(data), "\n") {
-		fields := strings.Fields(line)
-		if len(fields) < 2 {
-			continue
-		}
-		v, _ := strconv.ParseUint(fields[1], 10, 64)
-		switch fields[0] {
-		case "MemTotal:":
-			total = v * 1024
-		case "MemAvailable:":
-			available = v * 1024
-		}
-	}
-	return
-}
-
-func readNetworkInfo() (bytesSent, bytesRecv uint64, err error) {
-	data, err := os.ReadFile("/proc/net/dev")
-	if err != nil {
-		return 0, 0, err
-	}
-	// Try eth0 or wlan0 first
-	for _, line := range strings.Split(string(data), "\n") {
-		trimmed := strings.TrimSpace(line)
-		if !strings.HasPrefix(trimmed, "eth0:") && !strings.HasPrefix(trimmed, "wlan0:") {
-			continue
-		}
-		parts := strings.SplitN(trimmed, ":", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		fields := strings.Fields(parts[1])
-		if len(fields) < 10 {
-			continue
-		}
-		bytesRecv, _ = strconv.ParseUint(fields[0], 10, 64)
-		bytesSent, _ = strconv.ParseUint(fields[8], 10, 64)
-		return
-	}
-	// Fallback: sum all interfaces
-	for _, line := range strings.Split(string(data), "\n") {
-		trimmed := strings.TrimSpace(line)
-		if !strings.Contains(trimmed, ":") {
-			continue
-		}
-		parts := strings.SplitN(trimmed, ":", 2)
-		if len(parts) < 2 {
-			continue
-		}
-		fields := strings.Fields(parts[1])
-		if len(fields) < 10 {
-			continue
-		}
-		r, _ := strconv.ParseUint(fields[0], 10, 64)
-		s, _ := strconv.ParseUint(fields[8], 10, 64)
-		bytesRecv += r
-		bytesSent += s
-	}
-	return
-}
-
-func readProcessRSS() uint64 {
-	data, err := os.ReadFile("/proc/self/statm")
-	if err != nil {
-		return 0
-	}
-	fields := strings.Fields(string(data))
-	if len(fields) < 2 {
-		return 0
-	}
-	rssPages, _ := strconv.ParseUint(fields[1], 10, 64)
-	return rssPages * uint64(os.Getpagesize())
-}
-
-func (h *Handler) handleSystemStats(w http.ResponseWriter, r *http.Request) {
-	cpuTotal, cpuIdle, _ := readCPURaw()
-	memTotal, memAvailable, _ := readMemoryInfo()
-	netSent, netRecv, _ := readNetworkInfo()
-	processRSS := readProcessRSS()
-
-	writeJSON(w, http.StatusOK, SystemStats{
-		CPU:       CPUStats{Total: cpuTotal, Idle: cpuIdle},
-		Memory:    MemoryStats{Total: memTotal, Available: memAvailable, ProcessRSS: processRSS},
-		Network:   NetworkStats{BytesSent: netSent, BytesRecv: netRecv},
-		Uptime:    formatUptime(time.Since(appStartTime)),
-		Timestamp: time.Now().Unix(),
-	})
-}
-
-// handleGetAutoDiscoverSettings returns the current auto_discover config. The
-// default_password is NEVER returned over the API (avoid plaintext leakage);
-// instead a boolean has_default_password indicates whether one is configured.
-func (h *Handler) handleGetAutoDiscoverSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-	ad := h.config.AutoDiscover
-	writeJSON(w, http.StatusOK, map[string]any{
-		"enabled":              ad.AutoDiscoverEnabled(),
-		"scan_interval":        ad.ScanIntervalSeconds,
-		"listen_for_hello":     ad.ListenForHelloEnabled(),
-		"network_interface":    ad.NetworkInterface,
-		"default_username":     ad.DefaultUsername,
-		"has_default_password": ad.DefaultPassword != "",
-		"ignore_scopes":        ad.IgnoreScopes,
-	})
-}
-
-// handleUpdateAutoDiscoverSettings updates the auto_discover config and persists
-// it to disk. All fields are optional (nil = unchanged); scan_interval is
-// floored to 30 to respect RPi-3B resource constraints.
-func (h *Handler) handleUpdateAutoDiscoverSettings(w http.ResponseWriter, r *http.Request) {
-	if h.config == nil {
-		WriteError(w, http.StatusInternalServerError, "config not available")
-		return
-	}
-	var body struct {
-		Enabled          *bool    `json:"enabled"`
-		ScanInterval     *int     `json:"scan_interval"`
-		ListenForHello   *bool    `json:"listen_for_hello"`
-		NetworkInterface *string  `json:"network_interface"`
-		DefaultUsername  *string  `json:"default_username"`
-		DefaultPassword  *string  `json:"default_password"`
-		IgnoreScopes     []string `json:"ignore_scopes"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		WriteError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-
-	ad := &h.config.AutoDiscover
-	if body.Enabled != nil {
-		ad.Enabled = body.Enabled
-	}
-	if body.ScanInterval != nil {
-		v := *body.ScanInterval
-		if v < 30 {
-			v = 30 // floor: RPi-3B resource constraint
-		}
-		ad.ScanIntervalSeconds = v
-	}
-	if body.ListenForHello != nil {
-		ad.ListenForHello = body.ListenForHello
-	}
-	if body.NetworkInterface != nil {
-		ad.NetworkInterface = *body.NetworkInterface
-	}
-	if body.DefaultUsername != nil {
-		ad.DefaultUsername = *body.DefaultUsername
-	}
-	// Only overwrite the password when the client explicitly sends a non-empty
-	// one. The GET handler never returns it, so the frontend sends it only when
-	// the user types a new value; an empty/nil is treated as "leave unchanged"
-	// so a save that doesn't touch the password field doesn't wipe it.
-	if body.DefaultPassword != nil && *body.DefaultPassword != "" {
-		ad.DefaultPassword = *body.DefaultPassword
-	}
-	if body.IgnoreScopes != nil {
-		ad.IgnoreScopes = body.IgnoreScopes
-	}
-
-	if err := config.Save(h.configPath, h.config); err != nil {
-		logger.Warn("failed to save config", "error", err)
-		WriteError(w, http.StatusInternalServerError, "failed to save config")
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]string{"status": "updated"})
-}
 
 // registerSystemRoutes registers system/stats/settings/backup/protocol/feature routes.
 func (h *Handler) registerSystemRoutes(r chi.Router) {
